+++ ui/src/dashboards/components/GaugeOptions.js
@@ -123,7 +123,7 @@ class GaugeOptions extends Component {
       )
 
       const isUnique = !colorsWithoutMinOrMax.some(
-        color => color.value === targetValue
+        color => color.value === targetValue && color.id !== threshold.id
       )
 
       allowedToUpdate = greaterThanMin && lessThanMax && isUnique
@@ -146,11 +146,11 @@ class GaugeOptions extends Component {
     handleUpdateAxes(newAxes)
   }
 
-  handleSortColors = () => {
-    const {gaugeColors, handleUpdateGaugeColors} = this.props
-    const sortedColors = _.sortBy(gaugeColors, color => color.value)
+  get sortedGaugeColors() {
+    const {gaugeColors} = this.props
+    const sortedColors = _.sortBy(gaugeColors, 'value')
 
-    handleUpdateGaugeColors(sortedColors)
+    return sortedColors
   }
 
   render() {
@@ -174,12 +174,10 @@ class GaugeOptions extends Component {
             >
               <span className="icon plus" /> Add Threshold
             </button>
-            {gaugeColors.map(color => (
+            {this.sortedGaugeColors.map((color, index) => (
               <Threshold
-                isMin={color.value === gaugeColors[0].value}
-                isMax={
-                  color.value === gaugeColors[gaugeColors.length - 1].value
-                }
+                isMin={index === 0}
+                isMax={index === gaugeColors.length - 1}
                 visualizationType="gauge"
                 threshold={color}
                 key={uuid.v4()}
@@ -188,7 +186,6 @@ class GaugeOptions extends Component {
                 onValidateColorValue={this.handleValidateColorValue}
                 onUpdateColorValue={this.handleUpdateColorValue}
                 onDeleteThreshold={this.handleDeleteThreshold}
-                onSortColors={this.handleSortColors}
               />
             ))}
           </div>